tests/wpt/web-platform-tests/css/css-position/parsing/position-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Positioned Layout Module Level 3: parsing position with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-position/#propdef-position">
<meta name="assert" content="position supports only the grammar 'static | relative | absolute | sticky | fixed'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("position", "auto");
test_invalid_value("position", "static relative");
</script>
</body>
</html>
